requirements.txt
streamlit
chromadb==0.4.15
PyMuPDF
python-dotenv
openai==1.54.0
httpx==0.27.0